comment_parser: Add Ruby support.
According to IANA [1] there is no official MIME type for Ruby code. I've
created the unofficial MIME 'text/x-ruby' for this purpose.

I've referenced the "Programming Ruby" guide [2] in creating this
package. It defines a comment as "[starting] with a # character and
run[ning] to the end of the line."

[1] https://www.iana.org/assignments/media-types/media-types.xhtml
[2] http://ruby-doc.com/docs/ProgrammingRuby

Closes #14.